some refactorings & change filenames
